requirements.txt
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
python-dateutil==2.8.1
requests==2.22.0
